Move the library code from auxin_cli to auxin to facilitate code re-use
