+++ auxin/Cargo.toml
@@ -60,5 +60,7 @@ mime_guess = "2.0.3"
 
 thiserror = "1.0"
 
+serde_arrays = "0.1.0"
+
 [dev-dependencies]
 uuid = { version = "0.8.2", features = ["v4"] }
