+++ auxin_cli/Cargo.toml
@@ -51,7 +51,6 @@ thiserror = "1.0"
 uuid = "0.8.2"
 
 reqwest = { version = "0.11.10", features = ["json"] }
-serde_arrays = "0.1.0"
 
 [dev-dependencies]
 terminator = "0.1.0"
